+++ packages/client/src/built-in/admin-layout/AdminLayout.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+import { InternalAdminLayout } from '.';
+import { AdminProvider } from './AdminProvider';
+import { NoticeArea } from './NoticeArea';
+
+export const AdminLayout = (props) => {
+  return (
+    <AdminProvider>
+      <NoticeArea />
+      <InternalAdminLayout {...props} />
+    </AdminProvider>
+  );
+};
