+++ packages/client/src/built-in/admin-layout/AdminLayoutPlugin.tsx
@@ -1,0 +1,12 @@
+import { Plugin } from '../../application';
+import { RemoteSchemaTemplateManagerPlugin } from '../../schema-templates';
+import { AdminLayout } from './AdminLayout';
+
+export class AdminLayoutPlugin extends Plugin {
+  async afterAdd() {
+    await this.app.pm.add(RemoteSchemaTemplateManagerPlugin);
+  }
+  async load() {
+    this.app.addComponents({ AdminLayout });
+  }
+}
